Added Unscheduled Fixed tasks.
Provide a way for adding tasks that require that takes
a fixed amount of time but must be done within a certain
period of time(hours). This gives the user flexibility
as now they do not just have to schedule a timing to
do a task, but just by keeping track of the number of
the number of hours they need so that they can plan
accordingly.

The Parser accepts a fixed task in the format of
fixed <description> /for <duration>, and the UI displays
the task as [F][] <description> (needs <hours> hours)

This task preserves the existing task model while
adding this fixed task function.