A-Assertions: added some assertions into the code.
Assertions that handle when Pickle is not successfully
loaded and index out of bound errors for tasklist was added.

This is to add periodic checks within the code that will
print out an error if the program crashes due to an
unexpected input by the user, or if Pickle is not
working as intended.